convert to callable app
