+++ isis/src/mro/apps/marci2isis/marci2isis.h
@@ -1,0 +1,11 @@
+#ifndef marci2isis_h
+#define marci2isis_h
+
+#include "Pvl.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void marci2isis(UserInterface &ui, Pvl *log);
+}
+
+#endif
